Create CategoryNeuron.py
CategoryNeuron created. Each neuron represents the complex unit circle and has a number of categories or classes. Each category can occur multiple times in the unit circle. This number is cat_periodicity.  Given a complex value, the category it falls into can be calculated. Given a category, the bisectors for each of its arcs can be calculated.